Adding updated Testcase-8
